Add Separate file for Image Service Tests, Add New Test Files
