+++ src/test/PDFTest/CrudPDFServiceTest.java
@@ -86,93 +86,7 @@ public class CrudPDFServiceTest {
   //    uploadTestAnnotatedFormPDF();
   //  }
 
-  @Test
-  public void uploadImageJPGToPDFTest() {
-    User user =
-      createUser()
-        .withUserType(UserType.Client)
-        .withUsername(username)
-        .withPasswordToHash(password)
-        .buildAndPersist(userDao);
-    TestUtils.login(username, password);
 
-    File file = new File(resourcesFolderPath + File.separator + "veteran-id-card-vic.jpg");
-    HttpResponse<String> uploadResponse =
-            Unirest.post(TestUtils.getServerUrl() + "/upload")
-                    .field("pdfType", "IDENTIFICATION_DOCUMENT")
-                    .header("Content-Disposition", "attachment")
-                    .field("file", file)
-                    .asString();
-
-    JSONObject uploadResponseJSON = TestUtils.responseStringToJSON(uploadResponse.getBody());
-    assertThat(uploadResponseJSON.getString("status")).isEqualTo("SUCCESS");
-  }
-
-  @Test
-  public void uploadImagePNGToPDFTest() {
-    User user =
-      createUser()
-        .withUserType(UserType.Client)
-        .withUsername(username)
-        .withPasswordToHash(password)
-        .buildAndPersist(userDao);
-    TestUtils.login(username, password);
-
-    File file = new File(resourcesFolderPath + File.separator + "CIS_401_Final_Progress_Report.pdf");
-    HttpResponse<String> uploadResponse =
-            Unirest.post(TestUtils.getServerUrl() + "/upload")
-                    .field("pdfType", "IDENTIFICATION_DOCUMENT")
-                    .header("Content-Disposition", "attachment")
-                    .field("file", file)
-                    .asString();
-
-    JSONObject uploadResponseJSON = TestUtils.responseStringToJSON(uploadResponse.getBody());
-    assertThat(uploadResponseJSON.getString("status")).isEqualTo("SUCCESS");
-  }
-
-  @Test
-  public void uploadImageToRotatedPDFTest() {
-    User user =
-      createUser()
-        .withUserType(UserType.Client)
-        .withUsername(username)
-        .withPasswordToHash(password)
-        .buildAndPersist(userDao);
-    TestUtils.login(username, password);
-
-    File file = new File(resourcesFolderPath + File.separator + "CIS_401_Final_Progress_Report.pdf");
-    HttpResponse<String> uploadResponse =
-            Unirest.post(TestUtils.getServerUrl() + "/upload")
-                    .field("pdfType", "IDENTIFICATION_DOCUMENT")
-                    .header("Content-Disposition", "attachment")
-                    .field("file", file)
-                    .asString();
-
-    JSONObject uploadResponseJSON = TestUtils.responseStringToJSON(uploadResponse.getBody());
-    assertThat(uploadResponseJSON.getString("status")).isEqualTo("SUCCESS");
-  }
-
-  @Test
-  public void uploadInvalidImageToPDFTest() {
-    User user =
-      createUser()
-        .withUserType(UserType.Client)
-        .withUsername(username)
-        .withPasswordToHash(password)
-        .buildAndPersist(userDao);
-    TestUtils.login(username, password);
-
-    File file = new File(resourcesFolderPath + File.separator + "CIS_401_Final_Progress_Report.pdf");
-    HttpResponse<String> uploadResponse =
-            Unirest.post(TestUtils.getServerUrl() + "/upload")
-                    .field("pdfType", "IDENTIFICATION_DOCUMENT")
-                    .header("Content-Disposition", "attachment")
-                    .field("file", file)
-                    .asString();
-
-    JSONObject uploadResponseJSON = TestUtils.responseStringToJSON(uploadResponse.getBody());
-    assertThat(uploadResponseJSON.getString("status")).isEqualTo("SUCCESS");
-  }
 
   @Test
   public void uploadValidPDFTestExists() {
